Reorganizing code structure and adding service layer.
