--- src/main/java/dev/vitorvidal/characterCreator/application/controller/HelloController.java
@@ -1,15 +1,0 @@
-package dev.vitorvidal.characterCreator.application.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/hello")
-public class HelloController {
-
-    @GetMapping
-    public void getAll() {
-        System.out.println("Hello");
-    }
-}
