+++ src/main/java/dev/vitorvidal/characterCreator/application/service/CharacterService.java
@@ -1,0 +1,7 @@
+package dev.vitorvidal.characterCreator.application.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class CharacterService {
+}
