+++ src/main/java/dev/vitorvidal/characterCreator/application/vo/AbilityScoreVO.java
@@ -1,0 +1,4 @@
+package dev.vitorvidal.characterCreator.application.vo;
+
+public class AbilityScoreVO {
+}
