+++ src/main/java/dev/vitorvidal/characterCreator/domain/model/Character.java
@@ -1,8 +1,4 @@
-package dev.vitorvidal.characterCreator.domain.character;
-
-import dev.vitorvidal.characterCreator.domain.ability.AbilityScore;
-import dev.vitorvidal.characterCreator.model.Class;
-import dev.vitorvidal.characterCreator.model.Race;
+package dev.vitorvidal.characterCreator.domain.model;
 
 import javax.persistence.*;
 
@@ -12,19 +8,18 @@ public class Character {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private long id;
 
-    private Race race;
-    private Class characterClass;
     private String name;
     private int age;
+    private Race race;
+    private Class characterClass;
 
-    @ManyToOne
-    @JoinColumn(name = "ability_score_id")
-    private AbilityScore abilityScore;
+//    @ManyToOne
+//    @JoinColumn(name = "ability_score_id")
+//    private AbilityScore abilityScore;
 
     public Character(Race race, Class characterClass, AbilityScore abilityScore, String name, int age) {
         this.race = race;
         this.characterClass = characterClass;
-        this.abilityScore = abilityScore;
         this.name = name;
         this.age = age;
     }
