+++ src/main/java/dev/vitorvidal/characterCreator/domain/repository/AbilityScoreRepository.java
@@ -1,5 +1,6 @@
-package dev.vitorvidal.characterCreator.domain.ability;
+package dev.vitorvidal.characterCreator.domain.repository;
 
+import dev.vitorvidal.characterCreator.domain.model.AbilityScore;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
