+++ src/main/java/dev/vitorvidal/characterCreator/domain/repository/CharacterRepository.java
@@ -1,5 +1,6 @@
-package dev.vitorvidal.characterCreator.domain.character;
+package dev.vitorvidal.characterCreator.domain.repository;
 
+import dev.vitorvidal.characterCreator.domain.model.Character;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
